Change accuracy_multi function to match datatype

+++ fastai/metrics.py
@@ -13,5 +13,6 @@ def accuracy_thresh(thresh):
     return lambda preds,targs: accuracy_multi(preds, targs, thresh)
 
 def accuracy_multi(preds, targs, thresh):
-    return ((preds>thresh)==targs).float().mean()
+    return ((preds>thresh).float()==targs).float().mean()
+
